mcp-agent/pyproject.toml
[project]
name = "mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.63.3",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.1",
]
